+++ libs/components/src/lib/dialog/dialog.template.ts
@@ -1,4 +1,4 @@
-import {html, when} from '@microsoft/fast-element';
+import {html, slotted, when} from '@microsoft/fast-element';
 import type {ViewTemplate} from '@microsoft/fast-element';
 import type {
 	ElementDefinitionContext,
@@ -10,11 +10,12 @@ import { Icon } from '../icon/icon';
 import { Button } from '../button/button';
 import type {Dialog} from './dialog';
 
-const getClasses = ({iconPlacement}: Dialog) => classNames(
+const getClasses = ({iconPlacement, bodySlottedContent, footerSlottedContent} : Dialog) => classNames(
 	'base',
 	[`icon-placement-${iconPlacement}`, Boolean(iconPlacement)],
+	['hide-body', !bodySlottedContent?.length],
+	['hide-footer', !footerSlottedContent?.length],
 );
-
 /**
  *
  */
@@ -97,7 +98,7 @@ export const DialogTemplate: (
 		>
 			<slot name="main">
 				<div class="main-wrapper">
-					<div class="header">
+					<div class="header ${x => x.subtitle ? 'border' : ''}">
 							<slot name="graphic">
 								${when(x => x.icon, icon(iconTag))}
 							</slot>
@@ -105,8 +106,12 @@ export const DialogTemplate: (
 							${when(x => x.subtitle, subtitle())}
 							${renderDismissButton(buttonTag)}
 					</div>
-					<slot name="body"></slot>
-					<slot name="footer"></slot>
+					<div class="body ${x => x.bodySlottedContent?.length ? '' : 'hide'} ${x => x.fullWidthBody? 'full-width' : ''}" >
+						<slot name="body" ${slotted('bodySlottedContent')}></slot>
+					</div>
+					<div class="footer ${x => x.footerSlottedContent?.length ? '' : 'hide'}">
+						<slot name="footer" ${slotted('footerSlottedContent')}></slot>
+					</div>
 				</div>
 			</slot>
 		</dialog>