+++ libs/components/src/lib/select/select.ts
@@ -1,13 +1,13 @@
 import { applyMixins, Select as FoundationSelect } from '@microsoft/fast-foundation';
-import { attr } from '@microsoft/fast-element';
+import {attr, Observable} from '@microsoft/fast-element';
 import type { Popup } from '../popup/popup';
 import { FormElement, formElements } from '../../shared/patterns';
 import { AffixIcon } from '../../shared/patterns';
 import type { Appearance, Shape } from '../enums';
 
 
-type SelectAppearance = Extract<Appearance, Appearance.Outlined | Appearance.Ghost>;
-type SelectShape = Extract<Shape, Shape.Rounded | Shape.Pill>;
+export type SelectAppearance = Extract<Appearance, Appearance.Outlined | Appearance.Ghost>;
+export type SelectShape = Extract<Shape, Shape.Rounded | Shape.Pill>;
 
 /**
  * Base class for select
@@ -26,6 +26,11 @@ export class Select extends FoundationSelect {
 	override connectedCallback() {
 		super.connectedCallback();
 		this._popup.anchor = this._anchor;
+	}
+
+	override get displayValue(): string {
+		Observable.track(this, 'displayValue');
+		return this.firstSelectedOption?.getAttribute('label') ?? this.firstSelectedOption?.text ?? '';
 	}
 }