+++ libs/components/CHANGELOG.md
@@ -4,6 +4,33 @@ This log was last generated on Thu, 26 Jan 2023 14:40:25 GMT and should not be m
 
 <!-- Start content -->
 
+## [3.4.0](https://github.com/Vonage/vivid-3/compare/vivid-v3.3.0...vivid-v3.4.0) (2023-03-15)
+
+
+### Features
+
+* **documentation:** live samples ([#1030](https://github.com/Vonage/vivid-3/issues/1030)) ([ed4a88c](https://github.com/Vonage/vivid-3/commit/ed4a88c369937e89d1b6dd3f0e67960363a6eb45))
+
+
+### Bug Fixes
+
+* **appearance:** duotone color change ([#1051](https://github.com/Vonage/vivid-3/issues/1051)) ([29f58fe](https://github.com/Vonage/vivid-3/commit/29f58fed4737e70b287e09e9b3ff0e7f932ded77))
+* **public-contract:** export public contract constant types ([#1066](https://github.com/Vonage/vivid-3/issues/1066)) ([5d30baa](https://github.com/Vonage/vivid-3/commit/5d30baa7e1403a02d01939b0d7a078e990e2433a))
+* **select:** options width and selected-value ellipsis ([#1049](https://github.com/Vonage/vivid-3/issues/1049)) ([91c1148](https://github.com/Vonage/vivid-3/commit/91c11483a74ff60b17bfb7c880c41154d5ca9c79))
+* **tabs:** style fixes ([#1072](https://github.com/Vonage/vivid-3/issues/1072)) ([d8d4660](https://github.com/Vonage/vivid-3/commit/d8d46609275127566ad7d0131920dd632e091d0d))
+
+## [3.3.0](https://github.com/Vonage/vivid-3/compare/vivid-v3.2.0...vivid-v3.3.0) (2023-03-08)
+
+
+### Features
+
+* **select:** add label handling to option component ([#1047](https://github.com/Vonage/vivid-3/issues/1047)) ([560f7cd](https://github.com/Vonage/vivid-3/commit/560f7cd8e34d8d1685c0fa08ae4cc29690b0e8d6))
+
+
+### Bug Fixes
+
+* **components:** circular dependency in tree-item ([#1053](https://github.com/Vonage/vivid-3/issues/1053)) ([72984b9](https://github.com/Vonage/vivid-3/commit/72984b99b1f2cc727da222fee67d52490d9acb49))
+
 ## [3.2.0](https://github.com/Vonage/vivid-3/compare/vivid-v3.1.3...vivid-v3.2.0) (2023-02-19)